Fix #480 ('use super::' with internal lexer)
I was surprised to see that the infrastructure was already there,
and in fact the correct function was even already in use!
Just this argument was wrong.

I found it surprising that this would be the case, so I did some sleuthing
into the commit history. The single line modified by this commit was written
in this commit: 531536f4529f0bcb1a6b0b554

This line was probably copied from the similar line that generates the
'uses' at the very top of the generated grammar file.

You can see the test added for the aforementioned commit here:

https://github.com/lalrpop/lalrpop/pull/223/commits/541cb4cef5f0287867f00de4af771745bb5a3642#diff-a6e59d678291fca1872fd2212d565242df94e2bf23428bc5135a3a8c7e7bc250

That test had a very simple import of the form `use super::Name;`.
This worked, but for the wrong reason; it was causing the internal
module to import `Name` indirectly via the `use super::Name` that was
generated in the outer module.  Anything of the form
`use super::module::Name` or `use super::Name as Renamed;' would fail,
however, as `module` and `Name` wouldn't exist at that location.